+++ ref.bib
@@ -63,7 +63,7 @@ journal = {SIAM Journal on Scientific Computing},
 volume = {27},
 number = {1},
 pages = {19-42},
-doi = {10.1137/04060593},
+doi = {10.1137/04060593X},
 year = {2005}
 }
 
@@ -317,7 +317,8 @@ keywords = {programming languages, type systems, virtual machines, assembly lang
   number={9},
   pages={938--940},
   year={2014},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi = {10.1038/nmeth.3038}
 }
 
 @article{stegle2015computational,
@@ -379,7 +380,8 @@ url = {https://igraph.org},
   number={1},
   pages={1--11},
   year={2017},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41467-017-02001-5}
 }
 
 @article{ernst2019staged,
@@ -390,7 +392,8 @@ url = {https://igraph.org},
   number={1},
   pages={1--20},
   year={2019},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41467-019-09182-1}
 }
 
 @article{bacher2020low,
@@ -469,7 +472,8 @@ month = jul,
 	number={1},
 	pages={1--12},
 	year={2018},
-	publisher={BioMed Central}
+	publisher={BioMed Central},
+    doi={10.1186/s13059-018-1486-1}
 }
 
 @misc{cirrocumulus,
@@ -524,7 +528,8 @@ url = {https://CRAN.R-project.org/package=shiny},
   number={4},
   pages={319--338},
   year={2007},
-  publisher={IOS Press}
+  publisher={IOS Press},
+  doi={10.3233/ida-2007-11402}
 }
 
 @misc{scran.js,
@@ -669,5 +674,6 @@ url = {https://CRAN.R-project.org/package=shiny},
    Year="2016",
    Volume="17",
    Pages="75",
-   Month="Apr"
+   Month="Apr",
+   doi="10.1186/s13059-016-0947-7"
 }
