### 2023-05-18 #5474 @LTLA
Thanks @Kevin-Mattheus-Moerman.

DOIs are mostly fixed. The "Orchestrating single-cell analysis with Bioconductor" citation is a bit tricky as it refers to a book that accompanies the publication of the same name, and the suggested DOI refers to the latter. The book itself doesn't have a DOI unfortunately, though I can ask the BioC folks if they provisioned one.

As for the pre-print: no, this JOSS submission is the only one, in terms of a peer-reviewed submission. There's a bit of a backstory here but the sanitized version is that I post to biorXiv so that people can read it, not as a prelude to journal submission. It also helps that I can stuff more details in the biorXiv one, as I had to cut a lot to meet JOSS guidelines.
