### 2023-05-21 6848818
Use the paper's DOI for the book as a placeholder.
