+++ ref.bib
@@ -15,7 +15,8 @@
   Author="Lun, A. T. L. and Amezquita, R. A. and Gottardo, R.  and Hicks, S. C.",
   Publisher="Bioconductor",
   Year="2020",
-  url={https://bioconductor.org/books/release/OSCA/}
+  url={https://bioconductor.org/books/release/OSCA/},
+  doi={10.1038/s41592-019-0654-x}
 }
 
 @misc{knncolle,
