### 2023-05-21 #5474 @LTLA
@Kevin-Mattheus-Moerman I've just used the suggested DOI for now (to the manuscript), but it's not quite right. I'll update it with the proper DOI to the book if it ever becomes available.
