### 2023-05-30 #5474 @Kevin-Mattheus-Moerman
@LTLA from the inspection report it looks like the core functionality of your work is encapsulated by 19601 lines of JavaScript. Can you please clarify the amount of Javascript that you wrote (rather than libraries you might have imported). Perhaps you can point to specific files that capture your contribution? Thanks.
