### 2023-05-30 #5474 @LTLA
@Kevin-Mattheus-Moerman

I don't believe that the source repository contains any vendored libraries (other than a 400-line Pong game that we put in to distract users while the results were computing). @jkanche would know better, but AFAIK, most of that JS was handwritten for this app. For example, [this file](https://github.com/kanaverse/kana/blob/master/src/components/ExploreMode/index.js) is already 1.4 kLoC, and the other files are probably around about the same size. I guess UI development is just verbose; better to have complex code than to have a complex user experience.

If it's any consolation, the code for the **kana** application is just the tip of the JS iceberg. Underneath **kana**, there is the [**bakana**](https://github.com/kanaverse/bakana) library that orchestrates the compute pipeline. And underneath that, there is [**scran.js**](https://github.com/kanaverse/scran.js), which performs the individual compute steps. And underneath _that_, there are a dozen C++ libraries that each probably warrant their own submissions.
