### 2023-06-01 #5474 @jkanche
that is correct,  except for the [game files here](https://github.com/openjournals/joss-reviews/issues/5474#issuecomment-1568821560), everything is handwritten.

app is based on npm, none of the libraries are vendored.
