+++ static/js/484.0f697b69.chunk.js
@@ -1,2 +1,0 @@
-!function(){"use strict";var t={1484:function(t,n,r){var e,i=r(5831),a=r(7891),o=r(498),u=r(3143),s={},f=!1,c={},l={},d=!1;function h(t,n){s.init.numberOfCells();var r=o.Z(t),e=s.init.clone();try{for(;e.iterations()<n;)if(i.DG(e,{runTime:r,maxIterations:n}),t){var a=e.extractCoordinates();u.b({type:"tsne_iter",x:a.x,y:a.y,iteration:e.iterations()},[a.x.buffer,a.y.buffer])}s.final=e.extractCoordinates()}finally{e.free()}}u.j((function(t){var n=t.data.id;d?u.b({id:n,type:"error",error:"t-SNE worker is dead and cannot process messages"}):"INIT"==t.data.cmd?(e=i.j2(t.data.scranOptions)).then((function(t){u.b({id:n,type:"init_worker",data:{status:"SUCCESS"}})})).catch((function(t){u.b({id:n,type:"error",error:t})})):"RUN"==t.data.cmd?e.then((function(r){var e;"neighbors"in t.data?(a.Yl(s.neighbors),s.neighbors=o.q(t.data.neighbors),e=!0):e=!1;var d={perplexity:t.data.params.perplexity};e||a.xi(d,c)?(a.Yl(s.init),s.init=i.xv(s.neighbors,{perplexity:d.perplexity}),c=d,f=!0):f=!1;var p={iterations:t.data.params.iterations};(f||a.xi(p,l))&&(h(t.data.params.animate,p.iterations),l=p),u.b({id:n,type:"tsne_run",data:{status:"SUCCESS"}})})).catch((function(t){u.b({id:n,type:"error",error:t})})):"RERUN"==t.data.cmd?e.then((function(t){h(!0,l.iterations),u.b({id:n,type:"tsne_rerun",data:{status:"SUCCESS"}})})).catch((function(t){u.b({id:n,type:"error",error:t})})):"FETCH"==t.data.cmd?e.then((function(t){var r={x:s.final.x.slice(),y:s.final.y.slice(),iterations:l.iterations},e=[r.x.buffer,r.y.buffer];u.b({id:n,type:"tsne_fetch",data:r},e)})).catch((function(t){u.b({id:n,type:"error",error:t})})):"KILL"==t.data.cmd?(d=!0,e.then((function(t){i.RA(),u.b({id:n,type:"tsne_killed",data:null})})).catch((function(t){u.b({id:n,type:"error",error:t})}))):u.b({id:n,type:"error",error:"unknown message type '"+JSON.stringify(t.data)+"'"})}))},5831:function(t,n,r){r.d(n,{DG:function(){return o.DG},RA:function(){return e.RA},RJ:function(){return i.RJ},Wf:function(){return i.Wf},gd:function(){return i.gd},j2:function(){return e.j2},qv:function(){return a.qv},xv:function(){return o.xv}});var e=r(8782),i=r(6635),a=(r(8311),r(6663),r(489),r(6126),r(118),r(5644),r(4292),r(5470),r(1196),r(5219),r(5448),r(9206),r(2945),r(7339),r(6757),r(4609),r(8420),r(845),r(9283),r(6974),r(1184),r(4812),r(6199),r(5159),r(7067)),o=(r(9273),r(4181));r(5319),r(2494),r(5696),r(2732),r(9249),r(3779),r(7039),r(5538),r(7644),r(6961),r(2306),r(4021),r(4755)},4181:function(t,n,r){r.d(n,{DG:function(){return m},xv:function(){
[…]
