### 2023-06-09 a018893
Switch to some nicer colors based on the scater palette. (#226)
* Generate some nicer colors.
* switch to the new color palette
* bump version
---------

Co-authored-by: Jayaram Kancherla <jayaram.kancherla@gmail.com>
