+++ src/components/Plots/SVGDimPlot.js
@@ -1,6 +1,6 @@
 import { getMinMax } from "./utils";
 
-export const SVGDimPlot = (color, embeddata) => {
+export const SVGDimPlot = (color, embeddata, plabels, pgradient) => {
   const { x, y } = embeddata;
 
   // translate to new coordinates
@@ -8,11 +8,12 @@ export const SVGDimPlot = (color, embeddata) => {
   const yMinMax = getMinMax(y);
 
   const new_min = 10,
-    new_max = 490;
+    new_max_x = 440,
+    new_max_y = 490;
 
   function translateX(val) {
     let nval =
-      ((val - xMinMax[0]) / (xMinMax[1] - xMinMax[0])) * (new_max - new_min) +
+      ((val - xMinMax[0]) / (xMinMax[1] - xMinMax[0])) * (new_max_x - new_min) +
       new_min;
 
     return String(nval);
@@ -20,7 +21,7 @@ export const SVGDimPlot = (color, embeddata) => {
 
   function translateY(val) {
     let nval =
-      ((val - yMinMax[0]) / (yMinMax[1] - yMinMax[0])) * (new_max - new_min) +
+      ((val - yMinMax[0]) / (yMinMax[1] - yMinMax[0])) * (new_max_y - new_min) +
       new_min;
 
     return String(nval);
@@ -31,13 +32,82 @@ export const SVGDimPlot = (color, embeddata) => {
     circles.push(
       `<circle cx="${translateX(x)}" cy="${translateY(
         y[i]
-      )}" r="2" style="fill: ${color[i]}"></circle>`
+      )}" r="2" style="fill: ${color[i]}" />`
     );
   });
 
+  let legend = [];
+  let defs = [];
+  if (pgradient) {
+    const { factors, slider } = pgradient;
+
+    // defs.push(`linear-gradient(to right, #edc775 ${
+    //   ((slider[0] - factors[0]) * 100) / (factors[1] - factors[0])
+    // }%, #e09351, #df7e66, #b75347, #6d2f20 ${
+    //   100 - ((factors[1] - slider[1]) * 100) / (factors[1] - factors[0])
+    // }%)`);
+
+    const minPerc =
+        ((slider[0] - factors[0]) * 100) / (factors[1] - factors[0]),
+      maxPerc =
+        100 - ((factors[1] - slider[1]) * 100) / (factors[1] - factors[0]);
+
+    const binsSize = Math.round((maxPerc - minPerc) / 4);
+
+    defs.push(`
+      <linearGradient id="grad" gradientTransform="rotate(90)">
+        <stop offset="${minPerc}%" stop-color="#edc775" />
+        <stop offset="${minPerc + binsSize}%" stop-color="#e09351" />
+        <stop offset="${minPerc + 2 * binsSize}%" stop-color="#df7e66" />
+        <stop offset="${minPerc + 3 * binsSize}%" stop-color="#b75347" />
+        <stop offset="100%" stop-color="#6d2f20" />
+      </linearGradient>
+    `);
+
+    legend.push(
+      `<rect x="460" y="100" width="35" height="150" 
+      style="fill: url(#grad)" />`
+    );
+
+    legend.push(`<text x="${460}" y="${90}">${Math.round(factors[0])}</text>`);
+
+    if (slider[0] !== factors[0]) {
+      legend.push(
+        `<text x="${500}" y="${
+          90 + ((minPerc + binsSize) * 150) / 100
+        }">custom min: ${Math.round(slider[0])}</text>`
+      );
+    }
+
+    legend.push(`<text x="${460}" y="${270}">${Math.round(factors[1])}</text>`);
+
+    if (slider[1] !== factors[1]) {
+      legend.push(
[…]
