### 2023-06-09 d90fa51
fix file names, update gradients both categorical and gradient selections
