### 2023-06-09 d726f75
flip y coordinates, flip gradient bar and other minor changes
