### 2023-06-09 c68485c
combine styles for labels
