+++ src/components/Plots/SVGDimPlot.js
@@ -1,3 +1,4 @@
+import { max } from "d3";
 import { getMinMax } from "./utils";
 
 export const SVGDimPlot = (color, embeddata, plabels, pgradient) => {
@@ -63,7 +64,7 @@ export const SVGDimPlot = (color, embeddata, plabels, pgradient) => {
           <stop offset="${minPerc + binsSize}%" stop-color="#b75347" />
           <stop offset="${minPerc + 2 * binsSize}%" stop-color="#df7e66" />
           <stop offset="${minPerc + 3 * binsSize}%" stop-color="#e09351" />
-          <stop offset="${minPerc + 4 * binsSize}%" stop-color="#edc775" />
+          <stop offset="${maxPerc}%" stop-color="#edc775" />
           <stop offset="100%" stop-color="#edc775" />
         </linearGradient>
       `);
@@ -74,28 +75,26 @@ export const SVGDimPlot = (color, embeddata, plabels, pgradient) => {
       );
 
       legend.push(
-        `<text x="${475}" y="${270}">${Math.round(factors[0])}</text>`
+        `<text x="${475}" y="${270}">${Math.round(slider[0])}</text>`
       );
 
-      if (slider[0] !== factors[0]) {
-        legend.push(
-          `<text x="${500}" y="${
-            90 + ((minPerc + 4 * binsSize) * 150) / 100
-          }">custom min: ${Math.round(slider[0])}</text>`
-        );
-      }
+      // if (slider[0] !== factors[0]) {
+      //   legend.push(
+      //     `<text x="${500}" y="${
+      //       90 + ((100 - minPerc) * 150) / 100
+      //     }">custom min: ${Math.round(slider[0])}</text>`
+      //   );
+      // }
 
-      legend.push(
-        `<text x="${475}" y="${90}">${Math.round(factors[1])}</text>`
-      );
+      legend.push(`<text x="${475}" y="${90}">${Math.round(slider[1])}</text>`);
 
-      if (slider[1] !== factors[1]) {
-        legend.push(
-          `<text x="${500}" y="${
-            90 + (minPerc * 150) / 100
-          }">custom max: ${Math.round(slider[1])}</text>`
-        );
-      }
+      // if (slider[1] !== factors[1]) {
+      //   legend.push(
+      //     `<text x="${500}" y="${
+      //       90 + ((100 - maxPerc) * 150) / 100
+      //     }">custom max: ${Math.round(slider[1])}</text>`
+      //   );
+      // }
     } else {
       // const binsSize = Math.round((maxPerc - minPerc) / 2);
 
@@ -115,35 +114,35 @@ export const SVGDimPlot = (color, embeddata, plabels, pgradient) => {
 
       legend.push(
         `<text x="${475}" y="${270}" style="font-family: sans-serif">${Math.round(
-          factors[0]
+          slider[0]
         )}</text>`
       );
 
-      if (slider[0] !== factors[0]) {
-        legend.push(
-          `<text x="${500}" y="${
-            270 - (minPerc * 150) / 100
-          }" style="font-family: sans-serif">custom min: ${Math.round(
-            slider[0]
-          )}</text>`
-        );
-      }
+      // if (slider[0] !== factors[0]) {
+      //   legend.push(
+      //     `<text x="${500}" y="${
+      //       90 + ((100 - minPerc) * 150) / 100
+      //     }" style="font-family: sans-serif">custom min: ${Math.round(
+      //       slider[0]
+      //     )}</text>`
+      //   );
+      // }
 
       legend.push(
         `<text x="${475}" y="${90}" style="font-family: sans-serif">${Math.round(
-          factors[1]
+          slider[1]
         )}</text>`
       );
 
-      if (slider[1] !== factors[1]) {
[…]
