### 2023-06-10 dc70cb4
save dimensionality plots as SVG (#225)
* save dimensionality plots as SVG
* bump app version
