+++ src/components/Plots/DimPlot.js
@@ -27,6 +27,8 @@ import { AppToaster } from "../../AppToaster";
 import WebGLVis from "epiviz.gl";
 import { code } from "../../utils/utils";
 import { generateColors } from "./colors";
+import { SVGDimPlot } from "./SVGDimPlot";
+import { get_image_title } from "../Gallery/utils";
 
 const DimPlot = (props) => {
   const container = useRef();
@@ -671,11 +673,11 @@ const DimPlot = (props) => {
     if (containerEl) {
       // const iData = scatterplot.canvas.toDataURL();
 
-      let tmp = [...props?.savedPlot];
-
-      tmp.push({
+      let config = {
         color: cellColorArray,
+        coords: props?.redDimsData[props?.selectedRedDim],
         config: {
+          colorArray: cellColorArray,
           embedding: JSON.parse(JSON.stringify(props?.selectedRedDim)),
           annotation: JSON.parse(JSON.stringify(props?.colorByAnnotation)),
           highlight: plotGroups[props?.clusHighlight]
@@ -690,9 +692,122 @@ const DimPlot = (props) => {
             : null,
           geneIdx: props?.gene,
         },
-      });
+      };
+
+      if (showGradient) {
+        if (
+          props?.selectedFsetIndex !== null &&
+          props?.selectedFsetIndex !== undefined
+        ) {
+          config["gradient"] = {
+            factors: exprMinMax,
+            slider: sliderMinMax,
+            type: "single",
+          };
+        } else if (props?.gene !== null && props?.gene !== undefined) {
+          config["gradient"] = {
+            factors: exprMinMax,
+            slider: sliderMinMax,
+            type: "single",
+          };
+        }
+      } else {
+        if (!toggleFactorsGradient) {
+          config["gradient"] = {
+            factors: factorsMinMax,
+            slider: sliderFactorsMinMax,
+            type: "multi",
+          };
+        } else {
+          config["labels"] = {
+            labels: plotGroups,
+            colors: plotColorMappings,
+          };
+        }
+      }
+
+      let tmp = [...props?.savedPlot];
+      tmp.push(config);
 
       props?.setSavedPlot(tmp);
+    }
+  }
+
+  function handleSavePNG() {
+    const containerEl = container.current;
+    if (containerEl) {
+      // const iData = scatterplot.canvas.toDataURL();
+
+      let config = {
+        color: cellColorArray,
+        coords: props?.redDimsData[props?.selectedRedDim],
+        config: {
+          colorArray: cellColorArray,
+          embedding: JSON.parse(JSON.stringify(props?.selectedRedDim)),
+          annotation: JSON.parse(JSON.stringify(props?.colorByAnnotation)),
+          highlight: plotGroups[props?.clusHighlight]
+            ? JSON.parse(JSON.stringify(plotGroups[props?.clusHighlight]))
+            : plotGroups[props?.clusHighlight],
+          gene: props?.gene
+            ? JSON.parse(
+                JSON.stringify(
+                  genesInfo[geneColSel[props?.selectedModality]][props?.gene]
+                )
+              )
+            : null,
+          geneIdx: props?.gene,
+        },
+      };
+
+      if (showGradient) {
+        if (
+          props?.selectedFsetIndex !== null &&
+          props?.selectedFsetIndex !== undefined
+        ) {
+          config["gradient"] = {
+            factors: exprMinMax,
+            slider: sliderMinMax,
+            type: "single",
+          };
+        } else if (props?.gene !== null && props?.gene !== undefined) {
+          config["gradient"] = {
+            factors: exprMinMax,
+            slider: sliderMinMax,
+            type: "single",
+          };
+        }
+      } else {
+        if (!toggleFactorsGradient) {
+          config["gradient"] = {
[…]
