### 2023-06-14 3c4fa46
fix issue with reduced dims in gallery for explore mode, bump app version
