+++ static/js/612.079fcbc9.chunk.js
@@ -1,0 +1,2 @@
+!function(){"use strict";var e={9612:function(e,t,n){var r,s=n(4165),a=n(5861),o=n(3324),c=n(7762),u=n(797),i=n(5831),l=null;function f(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null;null===t&&(t=r.result.transaction(["analysis_meta"],"readonly").objectStore("analysis_meta"));var n=t.getAll();n.onsuccess=function(){var t=n.result;t.forEach((function(e){delete e.files})),e(t)},n.onerror=function(){e(null)}}function p(e,t){return h.apply(this,arguments)}function h(){return(h=(0,a.Z)((0,s.Z)().mark((function e(t,n){return(0,s.Z)().wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.abrupt("return",new Promise((function(e){var r=n.get(t);r.onsuccess=function(){void 0!==r.result?e(r.result):e(null)},r.onerror=function(){e(null)}})));case 1:case"end":return e.stop()}}),e)})))).apply(this,arguments)}function m(e){return Promise.allSettled(e).then((function(e){var t,n=(0,c.Z)(e);try{for(n.s();!(t=n.n()).done;){if(!t.value)return!1}}catch(r){n.e(r)}finally{n.f()}return!0}))}function d(){return y.apply(this,arguments)}function y(){return(y=(0,a.Z)((0,s.Z)().mark((function e(){return(0,s.Z)().wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,l;case 2:return e.abrupt("return",new Promise((function(e){f(e)})));case 3:case"end":return e.stop()}}),e)})))).apply(this,arguments)}function v(e,t){return b.apply(this,arguments)}function b(){return(b=(0,a.Z)((0,s.Z)().mark((function e(t,n){var a,o,c,u,i,f,h;return(0,s.Z)().wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,l;case 2:return a=r.result.transaction(["file","file_meta"],"readwrite"),o=a.objectStore("file"),c=a.objectStore("file_meta"),e.next=7,p(t,c);case 7:return u=e.sent,i=null===u?0:u.count,i++,f=new Promise((function(e){var r=o.put({id:t,payload:n.buffer});r.onsuccess=function(t){e(!0)},r.onerror=function(t){e(!1)}})),h=new Promise((function(e){u.count=i;var t=c.put(u);t.onsuccess=function(t){e(!0)},t.onerror=function(t){e(!1)}})),e.abrupt("return",m([f,h]));case 13:case"end":return e.stop()}}),e)})))).apply(this,arguments)}function _(e,t,n,r){return x.apply(this,arguments)}function x(){return(x=(0,a.Z)((0,s.Z)().mark((function e(t,n,a,o){var c,u,i,p,h,d;return(0,s.Z)().wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,l;case 2:if(c=r.result.transaction(["analysis","analysis_meta"],"readwrite"),u=c.objectStore("analysis"),i=c.objectStore("analysis_meta"),null!=t){e.next=10;break}return e.next=8,new Promise((function(e){return f(e,i)}));case 8:p=e.sent,t=String(p.length);case 10:return h=new Promise((function(e){var r=u.put({id:t,payload:n.buffer});r.onsuccess=function(t){e(!0)},r.onerror=
[…]
