### 2023-06-20 #5474 @LTLA
Hi @AoifeHughes, I picked a few appropriate names from the list based on "Javascript + bioinformatics"

- Robaina
- kblin
- holtgrewe
- camillescott

Hopefully a few of them will be interested.

Would you like some suggestions for people outside this list?
