### 2023-06-29 #5603 @editorialbot
[REVIEW]: Powering single-cell analyses in the browser with WebAssembly
**Submitting author:** @LTLA (<a href="http://orcid.org/0000-0002-3564-4813">Aaron Lun</a>)
**Repository:** https://github.com/kanaverse/kana
**Branch with paper.md** (empty if default branch): submission
**Version:** v3.0.1
