### 2023-07-04 06073f3
Wait the IndexedDB transactions to actually complete.
