### 2023-07-04 ff94ca0
Only check for completion of writing transactions.
