### 2023-07-04 fbbf3f8
Only check for completion of writing transactions.
