+++ src/workers/KanaDBHandler.js
@@ -1,7 +1,7 @@
 var kanaDB;
 var init = null;
 
-function complete(trans) {
+function complete_write(trans) {
   // See comments in DownloadsDB about the validity of wrapping this in a Promise.
   return new Promise((resolve, reject) => {
     trans.oncomplete = (event) => {
@@ -115,7 +115,7 @@ export async function getRecords() {
 export async function saveFile(id, buffer) {
   await init;
   let trans = kanaDB.result.transaction(["file", "file_meta"], "readwrite");
-  let fin = complete(trans);
+  let fin = complete_write(trans);
   let file_store = trans.objectStore("file");
   let meta_store = trans.objectStore("file_meta");
 
@@ -160,7 +160,7 @@ export async function saveAnalysis(id, state, files, title) {
     ["analysis", "analysis_meta"],
     "readwrite"
   );
-  let fin = complete(trans);
+  let fin = complete_write(trans);
   let analysis_store = trans.objectStore("analysis");
   let meta_store = trans.objectStore("analysis_meta");
 
@@ -225,7 +225,7 @@ export async function loadAnalysis(id) {
 export async function removeFile(id) {
   await init;
   let trans = kanaDB.result.transaction(["file", "file_meta"], "readwrite");
-  let fin = complete(trans);
+  let fin = complete_write(trans);
   let file_store = trans.objectStore("file");
   let meta_store = trans.objectStore("file_meta");
 
@@ -285,7 +285,7 @@ export async function removeAnalysis(id) {
     ["analysis", "analysis_meta"],
     "readwrite"
   );
-  let fin = complete(trans);
+  let fin = complete_write(trans);
   let analysis_store = trans.objectStore("analysis");
   let meta_store = trans.objectStore("analysis_meta");
 
