### 2023-07-04 f01d637
Repaired the KanaDB callbacks for more clarity.
