+++ src/workers/KanaDBHandler.js
@@ -44,7 +44,7 @@ export function initialize() {
 
     // Send existing stored analyses, if available.
     kanaDB.onsuccess = () => {
-      getRecordsResolver(resolve);
+      resolve(getRecordsInternal());
     };
 
     kanaDB.onerror = () => {
@@ -55,60 +55,33 @@ export function initialize() {
   return init;
 }
 
-function getRecordsResolver(resolve, store = null) {
-  if (store === null) {
-    store = kanaDB.result
+function getRecordsInternal() {
+  let store = kanaDB.result
       .transaction(["analysis_meta"], "readonly")
       .objectStore("analysis_meta");
-  }
 
   var allAnalysis = store.getAll();
+  return new Promise(resolve => {
+    allAnalysis.onsuccess = event => {
+      let vals = allAnalysis.result;
 
-  allAnalysis.onsuccess = function () {
-    let vals = allAnalysis.result;
-    vals.forEach((x) => {
-      delete x.files;
-    }); // no need to transfer that.
-    resolve(vals);
-  };
-  allAnalysis.onerror = function () {
-    resolve(null);
-  };
-}
+      // no need to transfer the files themselves.
+      vals.forEach((x) => {
+        delete x.files;
+      }); 
 
-/** Helper functions **/
-async function loadContent(id, store) {
-  return new Promise((resolve) => {
-    let request = store.get(id);
-    request.onsuccess = function () {
-      if (request.result !== undefined) {
-        resolve(request.result);
-      } else {
-        resolve(null);
-      }
+      resolve(vals);
     };
-    request.onerror = function () {
-      resolve(null);
+  
+    allAnalysis.onerror = event => {
+      reject(new Error(`failed to query the analysis store in KanaDB: ${event.target.errorCode}`));
     };
-  });
-}
-
-function allOK(promises) {
-  return Promise.allSettled(promises).then((vals) => {
-    for (const x of vals) {
-      if (!x) {
-        return false;
-      }
-    }
-    return true;
   });
 }
 
 export async function getRecords() {
   await init;
-  return new Promise((resolve) => {
-    getRecordsResolver(resolve);
-  });
+  return getRecordsInternal();
 }
 
 /** Functions to save content **/
@@ -119,39 +92,47 @@ export async function saveFile(id, buffer) {
   let file_store = trans.objectStore("file");
   let meta_store = trans.objectStore("file_meta");
 
-  var meta = await loadContent(id, meta_store);
-  var refcount;
-  if (meta === null) {
-    refcount = 0;
-  } else {
-    refcount = meta["count"];
-  }
-  refcount++;
+  let request = store.get(id);
+  await (new Promise((resolve, reject) => {
+    request.onsuccess = event => {
+      let meta = request.result;
+      if (typeof meta === "undefined") {
+        meta = { count: 1, id: id };
+      } else {
+        meta.count++;
+      }
 
-  var data_saving = new Promise((resolve) => {
-    var putrequest = file_store.put({ id: id, payload: buffer.buffer });
-    putrequest.onsuccess = function (event) {
-      resolve(true);
+      var data_saving = new Promise((resolve) => {
+        var putrequest = file_store.put({ id: id, payload: buffer.buffer });
+        putrequest.onsuccess = event => {
+          resolve(true);
+        };
+        putrequest.onerror = event => {
+          reject(new Error(`failed to save file ${id} in KanaDB: ${event.target.errorCode}`));
+        };
+      });
+
+      var ref_saving = new Promise((resolve) => {
+        meta.count = refcount;
+        var putrequest = meta_store.put(meta);
+        putrequest.onsuccess = event => {
+          resolve(true);
+        };
+        putrequest.onerror = event => {
[…]
