### 2023-07-04 44bf23b
Got rid of the completion helper for greater callback clarity.
Ensure that awaits are only used when all event handlers have been attached.
