### 2023-07-05 21c3313
Further fixes to the KanaDB handler, updated the scran worker.
