+++ src/workers/KanaDBHandler.js
@@ -32,7 +32,7 @@ export function initialize() {
 
     // Send existing stored analyses, if available.
     kanaDB.onsuccess = () => {
-      resolve(getRecordsInternal());
+      resolve(get_records());
     };
 
     kanaDB.onerror = () => {
@@ -43,7 +43,7 @@ export function initialize() {
   return init;
 }
 
-function getRecordsInternal() {
+function get_records() {
   let store = kanaDB.result
       .transaction(["analysis_meta"], "readonly")
       .objectStore("analysis_meta");
@@ -69,10 +69,11 @@ function getRecordsInternal() {
 
 export async function getRecords() {
   await init;
-  return getRecordsInternal();
+  return get_records();
 }
 
 /** Functions to save content **/
+
 export async function saveFile(id, buffer) {
   await init;
   let trans = kanaDB.result.transaction(["file", "file_meta"], "readwrite");
@@ -98,7 +99,7 @@ export async function saveFile(id, buffer) {
         meta.count++;
       }
 
-      var data_saving = new Promise((resolve) => {
+      var data_saving = new Promise((resolve, reject) => {
         var putrequest = file_store.put({ id: id, payload: buffer.buffer });
         putrequest.onsuccess = event => {
           resolve(true);
@@ -108,7 +109,7 @@ export async function saveFile(id, buffer) {
         };
       });
 
-      var ref_saving = new Promise((resolve) => {
+      var ref_saving = new Promise((resolve, reject) => {
         var putrequest = meta_store.put(meta);
         putrequest.onsuccess = event => {
           resolve(true);
@@ -176,7 +177,8 @@ export async function saveAnalysis(id, state, files, title) {
       };
     });
 
-    return Promise.all([data_saving, id_saving]).then(x => new_id);
+    // DO NOT await the promises here!
+    return [new_id, data_saving, id_saving];
   };
 
   let output_promise;
@@ -197,10 +199,11 @@ export async function saveAnalysis(id, state, files, title) {
   // Only await after attaching all event handlers.
   let output = await output_promise;
   await fin;
-  return output;
+  return output[0];
 }
 
 /** Functions to load content **/
+
 export async function loadFile(id) {
   await init;
   let file_store = kanaDB.result
@@ -219,10 +222,9 @@ export async function loadFile(id) {
 
   var meta = await meta_promise;
   if (meta !== null) {
-    return new Uint8Array(meta["payload"]);
-  } else {
-    return null;
+    meta = new Uint8Array(meta["payload"]);
   }
+  return meta;
 }
 
 export async function loadAnalysis(id) {
@@ -243,13 +245,84 @@ export async function loadAnalysis(id) {
 
   var meta = await ana_promise;
   if (meta !== null) {
-    return new Uint8Array(meta["payload"]);
-  } else {
-    return null;
+    meta = new Uint8Array(meta["payload"]);
   }
+  return meta;
 }
 
-/** Functions to load content **/
+/** Functions to remove content **/
+
+async function superResolver(x) {
+  let resolved = await x;
+  if (resolved instanceof Array) {
+    let replacement = [];
+    for (const y of resolved) {
+      replacement.push(await superResolver(y));
+    }
+    resolved = replacement;
+  }
+  return resolved;
+}
+
+function remove_file(id, file_store, meta_store) {
+  let request = meta_store.get(id);
+
+  return new Promise((resolve, reject) => {
+    request.onsuccess = event => {
+      let meta = request.result;
+      var refcount = meta["count"] - 1;
+      var promises = [];
+
+      if (refcount === 0) {
+        promises.push(
+          new Promise((resolve, reject) => {
+            let request = file_store.delete(id);
+            request.onsuccess = event => {
+              resolve(true);
+            };
+            request.onerror = event => {
[…]
