### 2023-07-06 9091983
Wait for transactions to actually complete in IndexedDB write operations. (#227)
* Wait the IndexedDB transactions to actually complete
* fixes to the KanaDB handler, updated the scran worker
* Ensure that awaits are only used when all event handlers have been attached
