+++ static/js/308.88fcb1e4.chunk.js
@@ -1,0 +1,2 @@
+"use strict";(self.webpackChunkkana=self.webpackChunkkana||[]).push([[308],{1106:function(e,r,t){e.exports=t.p+"static/media/scran.ff975bdaf7a33f6eab10.wasm"},8308:function(e,r,t){t.r(r);var n=function(){var e="file:///home/runner/work/kana/kana/node_modules/scran.js/browser/wasm/scran.js";return function(r){function n(){return F.buffer!=U&&ie(F.buffer),L}function o(){return F.buffer!=U&&ie(F.buffer),H}function i(){return F.buffer!=U&&ie(F.buffer),Y}function a(){return F.buffer!=U&&ie(F.buffer),q}function u(){return F.buffer!=U&&ie(F.buffer),V}function s(){return F.buffer!=U&&ie(F.buffer),X}function c(){return F.buffer!=U&&ie(F.buffer),K}var l,f,d="undefined"!=typeof(r=r||{})?r:{};d.ready=new Promise((function(e,r){l=e,f=r}));var p,h,m,y=Object.assign({},d),v=[],g="./this.program",w=function(e,r){throw r},_="object"==typeof window,E="function"==typeof importScripts,b="object"==typeof process&&"object"==typeof process.versions&&"string"==typeof process.versions.node,k=d.ENVIRONMENT_IS_PTHREAD||!1,C="";function j(e){return d.locateFile?d.locateFile(e,C):C+e}(_||E)&&(E?C=self.location.href:"undefined"!=typeof document&&document.currentScript&&(C=document.currentScript.src),e&&(C=e),C=0!==C.indexOf("blob:")?C.substr(0,C.replace(/[?#].*/,"").lastIndexOf("/")+1):"",p=function(e){var r=new XMLHttpRequest;return r.open("GET",e,!1),r.send(null),r.responseText},E&&(m=function(e){var r=new XMLHttpRequest;return r.open("GET",e,!1),r.responseType="arraybuffer",r.send(null),new Uint8Array(r.response)}),h=function(e,r,t){var n=new XMLHttpRequest;n.open("GET",e,!0),n.responseType="arraybuffer",n.onload=function(){200==n.status||0==n.status&&n.response?r(n.response):t()},n.onerror=t,n.send(null)});var T=d.print||console.log.bind(console),P=d.printErr||console.warn.bind(console);function S(e){S.shown||(S.shown={}),S.shown[e]||(S.shown[e]=1,P(e))}Object.assign(d,y),y=null,d.arguments&&(v=d.arguments),d.thisProgram&&(g=d.thisProgram),d.quit&&(w=d.quit);var D;Atomics.load,Atomics.store,Atomics.compareExchange;d.wasmBinary&&(D=d.wasmBinary);var F,A,M=d.noExitRuntime||!0;"object"!=typeof WebAssembly&&ge("no native wasm support detected");var $,O=!1;function x(e,r){e||ge(r)}var R="undefined"!=typeof TextDecoder?new TextDecoder("utf8"):void 0;function W(e,r,t){for(var n=(r>>>=0)+t,o=r;e[o]&&!(o>=n);)++o;if(o-r>16&&e.buffer&&R)return R.decode(e.buffer instanceof SharedArrayBuffer?e.slice(r,o):e.subarray(r,o));for(var i="";r<o;){var a=e[r++];if(128&a){var u=63&e[r++];if(192!=(224&a)){var s=63&e[r++];if((a=224==(
[…]

[…]
