+++ static/js/797.b08401bb.chunk.js
@@ -1,3 +1,0 @@
-/*! For license information please see 797.b08401bb.chunk.js.LICENSE.txt */
-"use strict";(self.webpackChunkkana=self.webpackChunkkana||[]).push([[797,463],{8109:function(e,t,n){n.d(t,{FB:function(){return F}});var r=n(4165),a=n(5671),i=n(3144);function s(e,t,n,r){return new(n||(n=Promise))((function(a,i){function s(e){try{u(r.next(e))}catch(t){i(t)}}function o(e){try{u(r.throw(e))}catch(t){i(t)}}function u(e){var t;e.done?a(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(s,o)}u((r=r.apply(e,t||[])).next())}))}var o,u=function(){function e(){(0,a.Z)(this,e),this.mutex=Promise.resolve()}return(0,i.Z)(e,[{key:"lock",value:function(){var e=function(){};return this.mutex=this.mutex.then((function(){return new Promise(e)})),new Promise((function(t){e=t}))}},{key:"dispatch",value:function(e){return s(this,void 0,void 0,(0,r.Z)().mark((function t(){var n;return(0,r.Z)().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,this.lock();case 2:return n=t.sent,t.prev=3,t.next=6,Promise.resolve(e());case 6:return t.abrupt("return",t.sent);case 7:return t.prev=7,n(),t.finish(7);case 10:case"end":return t.stop()}}),t,this,[[3,,7,10]])})))}}]),e}();var l="undefined"!==typeof globalThis?globalThis:"undefined"!==typeof self?self:"undefined"!==typeof window?window:n.g,c=null!==(o=l.Buffer)&&void 0!==o?o:null,f=l.TextEncoder?new l.TextEncoder:null;function h(e,t){return(15&e)+(e>>6|e>>3&8)<<4|(15&t)+(t>>6|t>>3&8)}function d(e,t){for(var n=t.length>>1,r=0;r<n;r++){var a=r<<1;e[r]=h(t.charCodeAt(a),t.charCodeAt(a+1))}}function p(e,t){if(e.length!==2*t.length)return!1;for(var n=0;n<t.length;n++){var r=n<<1;if(t[n]!==h(e.charCodeAt(r),e.charCodeAt(r+1)))return!1}return!0}var v="a".charCodeAt(0)-10,_="0".charCodeAt(0);function m(e,t,n){for(var r=0,a=0;a<n;a++){var i=t[a]>>>4;e[r++]=i>9?i+v:i+_,i=15&t[a],e[r++]=i>9?i+v:i+_}return String.fromCharCode.apply(null,e)}for(var w=null!==c?function(e){if("string"===typeof e){var t=c.from(e,"utf8");return new Uint8Array(t.buffer,t.byteOffset,t.length)}if(c.isBuffer(e))return new Uint8Array(e.buffer,e.byteOffset,e.length);if(ArrayBuffer.isView(e))return new Uint8Array(e.buffer,e.byteOffset,e.byteLength);throw new Error("Invali
[…]
