### 2023-07-19 bb7e891
Add docker scripts for building static site
Adds a couple of docker scripts to allow the static site to be easily
built, or to allow the site to be run locally for development.
