+++ docker-compose.yml
@@ -1,0 +1,12 @@
+services:
+  web:
+    build:
+      context: .
+      args:
+        UID: 1001
+        GID: 1001
+    ports:
+      - "3000:3000"
+    volumes:
+      - .:/home/kana
+
