### 2023-07-19 kanaverse/kana#231 @yongrenjie
[JOSS] Caption error
(This issue is opened as part of @yongrenjie and @llewelld's review of the Journal of Open Source Software submission, [*'Powering single-cell analyses in the browser with WebAssembly'*](https://github.com/openjournals/joss-reviews/issues/5603).)

---

In [the paper](https://github.com/kanaverse/kana/blob/6848818ce21fefa80078227d3e3e803f2608efd2/paper.md?plain=1#L72-L74), the caption to Figure 2 appears to have an error: "the bottom-left panel contains a gallery of miscellaneous plots", it looks like should be the "panel on the right hand edge".
