### 2023-07-19 kanaverse/kana#232 @yongrenjie
[JOSS] Expand on performance claims
(This issue is opened as part of @yongrenjie and @llewelld's review of the Journal of Open Source Software submission, [*'Powering single-cell analyses in the browser with WebAssembly'*](https://github.com/openjournals/joss-reviews/issues/5603).)

---

There is a [single performance claim](https://github.com/kanaverse/kana/blob/6848818ce21fefa80078227d3e3e803f2608efd2/paper.md?plain=1#L100C10-L100C10) that for a 170 000-cell dataset kana finished its analysis in under 5 minutes and using less than 3GB RAM on a consumer laptop. It would help greatly to give details of the set up (laptop configuration, browser, etc.) and also to compare against the non-WASM native-compiled version of scran.
