### 2023-07-19 kanaverse/kana#233 @yongrenjie
[JOSS] Add description for non-specialist audience
(This issue is opened as part of @yongrenjie and @llewelld's review of the Journal of Open Source Software submission, [*'Powering single-cell analyses in the browser with WebAssembly'*](https://github.com/openjournals/joss-reviews/issues/5603).)

---

[The summary](https://github.com/kanaverse/kana/blob/6848818ce21fefa80078227d3e3e803f2608efd2/paper.md?plain=1#L23) gives a clear idea about the scientific area the tool works in and its particular benefits (those being 'omics and in-browser client-side operation). These are great points that the text is clear on. However, in terms of a clear description for a "diverse, non-specialist audience", we feel a little more detail would be helpful, in particular explaining what some of the consequences (scientific outputs) of using the software might allow, since this helps put the software in a context that's easier for non-specialists to grasp. For example: "This tool allows variations between cells in a population to be studied, in order to understand how biochemical diversity underlies and gives rise to physiological processes".
