### 2023-07-19 kanaverse/kana#234 @yongrenjie
[JOSS] Comparison of web app and CLI from user perspective
(This issue is opened as part of @yongrenjie and @llewelld's review of the Journal of Open Source Software submission, [*'Powering single-cell analyses in the browser with WebAssembly'*](https://github.com/openjournals/joss-reviews/issues/5603).)

---

The paper provides a cogent description of the benefits of running single-cell analyses in the browser from a *developer* perspective: for example, not having to provision a backend server. However, given that the main contribution of the tool is client-side visualisation and execution, it would be useful to have an idea about the benefits and drawbacks of this approach *for the user*. In particular:
