### 2023-07-19 kanaverse/kana#235 @yongrenjie
[JOSS] Comparisons with other apps 
(This issue is opened as part of @yongrenjie and @llewelld's review of the Journal of Open Source Software submission, [*'Powering single-cell analyses in the browser with WebAssembly'*](https://github.com/openjournals/joss-reviews/issues/5603).)

---

In terms of contextualising the software in relation to other work, there is discussion about other browser-based bioinformatics tools, but no discussion about alternative 'omics software for cell clustering and marker detection. It would be helpful to compare against other software in this area, especially those which are also browser-based, or (if nothing relevant exists) to emphasise the uniqueness of the tool.
