### 2023-07-19 kanaverse/kana#236 @yongrenjie
[JOSS] Community guidelines
(This issue is opened as part of @yongrenjie and @llewelld's review of the Journal of Open Source Software submission, [*'Powering single-cell analyses in the browser with WebAssembly'*](https://github.com/openjournals/joss-reviews/issues/5603).)

--

The project makes good use of GitHub Issues, and it was particularly encouraging to see that the app even points the user at the Issues pages from within the app's error dialogues. However, the guidelines for contribution are minimal (e.g. no code of conduct, just the basic instructions for getting a build server running). This would be a good area to improve upon. Clearly at present the authors have been the main contributors, and fostering a community around the tool would not only benefit development, but it would also seem that the complexity and level of professionalism of the software means there's potential for supporting a vibrant community around the project.
