### 2023-07-19 kanaverse/kana#237 @yongrenjie
[JOSS] Instructions for deployment
(This issue is opened as part of @yongrenjie and @llewelld's review of the Journal of Open Source Software submission, [*'Powering single-cell analyses in the browser with WebAssembly'*](https://github.com/openjournals/joss-reviews/issues/5603).)

See also: #230

----

In the [kana README.md](https://github.com/kanaverse/kana/tree/master#readme), it states that "Deployment is as easy as serving the static files in this repository via HTTPS." In practice the static files aren't immediately available in the repository and the react app has to be built. We did notice the files in the static [`kana-v3-dev` repository](https://github.com/kanaverse/kana-v3-dev), so that could be referred to as a way to get the built files. However, instructions for how to build them would be even better.

For example, the following worked for us:

```
yarn && PUBLIC_URL="/kana" yarn build
```
Then the files in the `build` folder can be transferred to a server for hosting.

We have set up a docker compose file that performs this, see the related PR (#230).

```
git clone https://github.com/kanaverse/kana.git
cd kana
docker compose up
```

Having built the app, we were able to deploy to a web server and have it run identically to the authors' own installation, with no server-side processing, which is superb, and validates the authors' claims.

There are instructions for contributing, which although brief, go some way to explaining what a developer would have to do to get a local system running (we used them ourselves to build the app). In its current form, however, this isn't quite enough to build the pages for transferring to a server: we had to do a bit more work than was described.
