### 2023-07-19 kanaverse/kana#238 @yongrenjie
[JOSS] Web app usage guidance
(This issue is opened as part of @yongrenjie and @llewelld's review of the Journal of Open Source Software submission, [*'Powering single-cell analyses in the browser with WebAssembly'*](https://github.com/openjournals/joss-reviews/issues/5603).)

----

Documentation is fairly sparse, in that there's a README.md, but no other documentation in the repo. However, the README.md does give a high-level overview with relevant content and points the user to a book, written by the authors (and others), which follows the same flow (although this uses the R version of scran).

When running, the app is well supported with instruction pages, tooltips and help buttons. However there isn't any detailed in-app reference documentation. The app is generally quite intuitive, so this isn't necessarily a problem, but a link to some more detailed documentation, or a tutorial document to walk the user through an example, would be helpful either in the README.md or as part of the app itself (there may already be something that we missed).
