### 2023-07-20 kanaverse/kana#231 @LTLA
Thanks, forgot to update it after we changed the UI layout. Should be fixed in kanaverse/kana-paper@0e6a58b.
