### 2023-07-20 kanaverse/kana#232 @LTLA
We have these timings in the [biorXiv version of the paper](https://www.biorxiv.org/content/10.1101/2022.03.02.482701v2) (see Table 2) but we had to cut them out to fit into JOSS's word limits. I'll copy it here:

> To evaluate the efficiency of our Wasm strategy, we compared a kana analysis in the browser to that of a native executable compiled from the same C++ libraries.  We analyzed several public scRNA-seq datasets (Table 1) using the default kana parameters for both approaches, i.e., QC filtering to 3 MADs from the median; PCA on the top 2500 HVGs to obtain the top 25 PCs; SNN graph construction with 10 neighbors and multi-level community detection at a resolution of 0.5; t-SNE with a perplexity of 30; UMAP with 15 neighbors and a minimum distance of 0.01; and 8 threads for all parallel sections (i.e., web workers for kana, see below).  We collected timings on an Intel Core i7-8850H CPU (2.60GHz, 6 cores, 32 GB memory) running Manjaro Linux. For convenience, we ran the kana timings in batch using Puppeteer to control a headless Chrome browser (HeadlessChrome/98.0.4758.0).
>
> Our results indicate that kana analyses took approximately 25-50% longer to run compared to the native executable (Table 2).  This is consistent with other benchmarking results (Jangda et al., 2019) where the performance gap is attributed to Wasm's design constraints and the overhead of the browser's Wasm runtime environment.  Our native executable was also created with a different compiler toolchain (GCC, instead of LLVM for the Wasm binary), where the same nominal optimization level (O3) may have different effects.  These results suggest that some work may still be required to completely fulfill Wasm's promise of "near-native execution".  Nonetheless, the current performance is largely satisfactory for kana, and will likely improve over time as browser implementations evolve along with our understanding of the relevant optimizations.
>
> | Dataset  | Number of cells | kana | Native |
> |----------|-----------------|------|--------|
> | Zeisel   | 3005            | 7.00 ± 0.10 | 5.60 ± 0.05 |
> | Paul     | 10368           | 17.59 ± 0.20 | 13.52 ± 0.38 |
> | Bach     | 25806           | 54.96 ± 1.13 | 43.33 ± 0.39 |
> | Ernst    | 68937           | 157.15 ± 7.39 | 114.67 ± 1.86 |
> | Bacher   | 104417          | 228.02 ± 2.85 | 170.32 ± 1.34 |
> | Zilionis | 173954          | 272.265 ± 4.22 | 183.77 ± 2.46 |

I can put all this stuff back in, though as you can see, there is a lot of associated commentary, e.g., description of the datasets, description of the laptop, description of the timing parameters and configuration, some discussion of the results. Probably would double the length of the current manuscript if I also added the text about memory usage as well. So I don't know whether JOSS (or more specifically, @AoifeHughes) would be willing to consider that.
