### 2023-07-21 kanaverse/kana#233 @LTLA
Added a few more sentences about what people usually try to get out of single-cell analyses in kanaverse/kana-paper#6.
