### 2023-07-21 c922ed4
Add docker scripts for building static site (#230)
Adds a couple of docker scripts to allow the static site to be easily
built, or to allow the site to be run locally for development.

Co-authored-by: David Llewellyn-Jones <david@flypig.co.uk>
