+++ static/js/933.6f14de8f.chunk.js
@@ -1,0 +1,2 @@
+!function(){"use strict";var n={1933:function(n,t,r){var e,i=r(5831),o=r(7891),a=r(498),u=r(3143),c={},f=!1,s={},d={},l=!1;function h(n){var t=a.Z(n),r=c.init.clone();try{for(c.total=r.totalEpochs();r.currentEpoch()<c.total;)if(i.q7(r,{runTime:t}),n){var e=r.extractCoordinates();u.b({type:"umap_iter",x:e.x,y:e.y,iteration:r.currentEpoch()},[e.x.buffer,e.y.buffer])}c.final=r.extractCoordinates()}finally{r.free()}}u.j((function(n){var t=n.data.id;l?u.b({id:t,type:"error",error:"UMAP worker is dead and cannot process messages"}):"INIT"==n.data.cmd?(e=i.j2(n.data.scranOptions)).then((function(n){u.b({id:t,type:"init_worker",data:{status:"SUCCESS"}})})).catch((function(n){u.b({id:t,type:"error",error:n})})):"RUN"==n.data.cmd?e.then((function(r){var e;"neighbors"in n.data?(o.Yl(c.neighbors),c.neighbors=a.q(n.data.neighbors),e=!0):e=!1;var l={min_dist:n.data.params.min_dist,num_epochs:n.data.params.num_epochs};e||o.xi(l,s)?(o.Yl(c.init),c.init=i.fD(c.neighbors,{epochs:l.num_epochs,minDist:l.min_dist}),s=l,f=!0):f=!1;var p={};(f||o.xi(p,d))&&(h(n.data.params.animate),d=p),u.b({id:t,type:"umap_run",data:{status:"SUCCESS"}})})).catch((function(n){u.b({id:t,type:"error",error:n})})):"RERUN"==n.data.cmd?e.then((function(n){h(!0),u.b({id:t,type:"umap_rerun",data:{status:"SUCCESS"}})})).catch((function(n){u.b({id:t,type:"error",error:n})})):"FETCH"==n.data.cmd?e.then((function(n){var r={x:c.final.x.slice(),y:c.final.y.slice(),iterations:c.total},e=[r.x.buffer,r.y.buffer];u.b({id:t,type:"umap_fetch",data:r},e)})).catch((function(n){u.b({id:t,type:"error",error:n})})):"KILL"==n.data.cmd?(l=!0,e.then((function(n){i.RA(),u.b({id:t,type:"umap_killed",data:null})})).catch((function(n){u.b({id:t,type:"error",error:n})}))):u.b({id:t,type:"error",error:"unknown message type '"+JSON.stringify(n.data)+"'"})}))},5831:function(n,t,r){r.d(t,{RA:function(){return e.RA},RJ:function(){return i.RJ},Wf:function(){return i.Wf},fD:function(){return a.fD},gd:function(){return i.gd},j2:function(){return e.j2},q7:function(){return a.q7},qv:function(){return o.qv}});var e=r(8782),i=r(6635),o=(r(8311),r(6663),r(489),r(6126),r(118),r(5644),r(4292),r(5470),r(1196),r(5219),r(5448),r(9206),r(2945),r(7339),r(6757),r(4609),r(8420),r(845),r(9283),r(6974),r(1184),r(4812),r(6199),r(5159),r(7067)),a=(r(9273),r(4181),r(5319));r(2494),r(5696),r(2732),r(9249),r(3779),r(7039),r(5538),r(7644),r(6961),r(2306),r(4021),r(4755)},4181:function(n,t,r){r(6635),r(8782),r(7108),r(7067)},5319:fun
[…]

[…]
