### 2023-07-21 ffa47e2
instructions to serve html locally
