### 2023-07-21 kanaverse/kana#237 @jkanche
I've merged David's PR and added instructions to the README for generating and serving the static HTML files. I believe there are many simpler ways other than docker compose to serve the html files.
