### 2023-07-21 kanaverse/kana#235 @LTLA
Added some comments in kanaverse/kana-paper#7.
