### 2023-07-21 8945dce
update README with contributing notes
