### 2023-07-21 kanaverse/kana#236 @jkanche
We've added CONTIRBUTING.md and updated the README
