+++ paper.md
@@ -16,7 +16,7 @@ authors:
 affiliations:
   - name: Genentech Inc., South San Francisco, USA
     index: 1
-date: 15 May 2023
+date: 20 July 2023
 bibliography: ref.bib
 ---
 
@@ -32,11 +32,15 @@ Results are presented in an intuitive web interface for further exploration and 
 
 # Statement of need
 
-Single-cell 'omics is routinely used to identify cell subpopulations or states from heterogeneous biological samples [@stegle2015computational].
-This typically involves several iterations of data analysis, visualization and interpretation by biologists who may not be familiar with programming frameworks for bioinformatics.
+Single-cell 'omics is routinely used to identify cell subpopulations with distinct molecular phenotypes in heterogeneous biological samples [@stegle2015computational].
+For example, in single cell RNA sequencing (scRNA-seq), cells are clustered based on their transcriptional profiles and each cluster is characterized based on differential expression of marker genes.
+Changes in expression or cellular abundance for each cluster can then provide some insight into the biological processes associated with the corresponding cell type or state.
+This analysis is often exploratory in nature as the properties of an interesting cell subpopulation are difficult to define _a priori_.
+As a result, each analysis involves several iterations of computation, visualization and interpretation by biologists who may not be familiar with programming frameworks for bioinformatics.
+
 Web applications offer an ideal environment for single-cell analyses, providing a user-friendly interface to the analysis workflow without any installation of additional software (other than a browser).
-Most existing web applications for single-cell analysis [@megill2021cellxgene;@cirrocumulus;@shiny] use a traditional server-based architecture,
-where data is sent to a backend server to compute results that are returned to the client (i.e., the user's machine) for visualization.
+This strategy has been used by tools such as Cellxgene [@megill2021cellxgene], Cirrocumulus [@cirrocumulus], and a variety of R/Shiny applications [@shiny] for processing single-cell data.
+The vast majority of these existing web applications use a traditional server-based architecture, where data is sent to a backend server to compute results that are returned to the client - i.e., the user's machine - for inspection.
 This obviously requires the deployment of a backend server, which has non-negligible cost when scaled to a large number of users.
 It also requires data transfer from the client to the server, which introduces latency as well as invoking concerns over data ownership and privacy.
 
@@ -68,14 +72,20 @@ The entire analysis can then be executed with a single click, though users can e
 Once each step of the analysis is complete, kana visualizes its results in a multi-panel layout (\autoref{screenshot:results}).
 One panel contains a scatter plot for the low-dimensional embeddings, where each cell is a point that is colored by cluster identity or gene expression.
 Another panel contains a table of marker statistics for a selected cluster, where potential marker genes are ranked and filtered according to the magnitude of upregulation over other clusters.
+Scientists can use this interface to examine the cellular heterogeneity in the dataset and to determine which biological processes are most active in each subpopulation.
 We also provide a gallery to visualize miscellaneous details such as the distribution of QC metrics.
 
-![Screenshot showing the multi-panel layout for results in the kana application. The top-left panel is used for the low-dimensional embeddings, the right panel contains the marker table for a selected cluster, and the bottom-left panel contains a gallery of miscellaneous plots.\label{screenshot:results}](screenshots/results.png)
+![Screenshot showing the multi-panel layout for results in the kana application. The left panel is used for the low-dimensional embeddings, the middle panel contains the marker table for a selected cluster, and the right panel contains a gallery of miscellaneous plots.\label{screenshot:results}](screenshots/results.png)
 
 Once the analysis is complete, users can export the analysis configuration and results for later inspection.
[…]
