### 2023-07-21 kanaverse/kana#234 @LTLA
I think your text got cut off.

Anyway, the "Further comments" sections has some comments about the consequences of client-side compute for the user. I've fleshed them out a little in kanaverse/kana-paper#8.
