### 2023-07-21 kanaverse/kana#232 @LTLA
In the meantime, I just added some brief details in kanaverse/kana-paper#9. Hopefully this is a satisfactory compromise.
