### 2023-07-25 ceb174f
Fix for default annotation color-by choice in explore mode.
