### 2023-07-25 ac8b932
Corrected display of CRISPR feature counts in analysis mode.
Generalized feature counts in explore mode for arbitrary modalities.
