+++ static/js/747.219694d0.chunk.js
@@ -1,2 +1,0 @@
-!function(){"use strict";var e={9747:function(e,n,t){var r,o=t(4165),s=t(5861),a=t(3324),c=t(7762),i=t(797),u=t(5831),l=null;function f(e){return p.apply(this,arguments)}function p(){return p=(0,s.Z)((0,o.Z)().mark((function e(n){var t,s,a,c,i,u,f,p,d,m,h,y,v,g=arguments;return(0,o.Z)().wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return t=g.length>1&&void 0!==g[1]?g[1]:null,s=g.length>2&&void 0!==g[2]&&g[2],e.next=4,l;case 4:if(s){e.next=13;break}return a=r.result.transaction(["downloads"],"readonly"),c=a.objectStore("downloads"),i=new Promise((function(e,t){var r=c.get(n);r.onsuccess=function(n){void 0!==r.result?e(r.result.payload):e(null)},r.onerror=function(e){t("failed to query DownloadsDB for ".concat(n,": ").concat(e.target.errorCode))}})),e.next=10,i;case 10:if(null===(u=e.sent)){e.next=13;break}return e.abrupt("return",u);case 13:return f=null==t?fetch(n):fetch(n,t),e.next=16,f;case 16:if((p=e.sent).ok){e.next=19;break}throw new Error("failed to download '"+n+"' ("+p.status+")");case 19:return e.next=21,p.arrayBuffer();case 21:return d=e.sent,m=r.result.transaction(["downloads"],"readwrite"),h=new Promise((function(e,t){m.oncomplete=function(n){e(null)},m.onerror=function(e){t(new Error("transaction error for saving ".concat(n," in DownloadsDB: ").concat(e.target.errorCode)))}})),y=m.objectStore("downloads"),v=new Promise((function(e,t){var r=y.put({url:n,payload:d});r.onsuccess=function(n){e(!0)},r.onerror=function(e){t(new Error("failed to cache ".concat(n," in DownloadsDB: ").concat(e.target.errorCode)))}})),e.next=28,v;case 28:return e.next=30,h;case 30:return e.abrupt("return",d);case 31:case"end":return e.stop()}}),e)}))),p.apply(this,arguments)}var d=t(9062),m=t(7325),h=t(105),y="https://cors-proxy.aaron-lun.workers.dev";function v(e){return g.apply(this,arguments)}function g(){return(g=(0,s.Z)((0,o.Z)().mark((function e(n){var t;return(0,o.Z)().wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,f(y+"/"+encodeURIComponent(n));case 2:return t=e.sent,e.abrupt("return",new Uint8Array(t));case 4:case"end":return e.stop()}}),e)})))).apply(this,arguments)}function w(e,n){if(e)if(Array.isArray(e)){var t,r=(0,c.Z)(e);try{for(r.s();!(t=r.n()).done;){w(t.value,n)}}catch(u){r.e(u)}finally{r.f()}}else if(e.constructor==Object)for(var o=0,s=Object.entries(e);o<s.length;o++){var i=(0,a.Z)(s[o],2);i[0];w(i[1],n)}else if(ArrayBuffer.isView(e)){if(!(e.buffer instanceof ArrayBuffer))throw"only ArrayBuffers should be in the message payload";n.push(e.buffer)}}function S(e){postMessage({type:"".concat(e,"_START")})}function b(e,n){if("undefined"=
[…]
