### 2023-07-31 kanaverse/kana#232 @llewelld
Thanks for providing the detail and for the changes to your text. For me this is really interesting material and seems relevant for JOSS; however I can also appreciate it's a problem to include it in full given the word limit.

I'm personally happy with the compromise wording. It gives a flavour of the results (although I think the summary you give in your bioRxiv paper is more useful in practice: "Our results indicate that kana analyses took approximately 25-50% longer to run compared to the native executable").

Is there a reason you don't cite your bioRxiv paper directly? This would seem like a natural thing to do and provide an easy way for the reader to drill down further into the detail.
