### 2023-07-31 kanaverse/kana#233 @llewelld
Thanks for the additions. These changes look good to me, but I'd defer to @yongrenjie for his opinion on this too.
