### 2023-07-31 kanaverse/kana#234 @llewelld
Apologies for the lost text; the full wording should have been as follows:

> The paper provides a cogent description of the benefits of running single-cell analyses in the browser from a *developer* perspective: for example, not having to provision a backend server. However, given that the main contribution of the tool is client-side visualisation and execution, it would be useful to have an idea about the benefits and drawbacks of this approach *for the user*. In particular:
>
>    1. What is the overhead of running it via WASM rather than natively or as an R package?
>    2. What order-of-magnitude size of data generally needed for single cell analysis, that would otherwise have to be transferred to the server?
>
> Biology is not our specific area of expertise, but we wonder whether, for a biologist working in bioinformatics, an R-based solution wouldn't nevertheless work well for them (we expect R to be a tool many biologists would be familiar with). Given this, would it be possible to more fully justify the use of the browser and the impacts of doing so in the text?

The changes you've already made in https://github.com/kanaverse/kana-paper/pull/8 and https://github.com/kanaverse/kana-paper/pull/9 already go a long way to addressing these comments in my view. Could you suggest representative dataset sizes in your "No latency from transfer of data" bullet point? If your benchmarking files are representative then I guess we're typically talking tens or hundreds of MB?
