### 2023-07-31 kanaverse/kana#236 @llewelld
This looks good to me. Thank you for adding this, I definitely think including this kind of material smooths the way for potential contributors.
