### 2023-07-31 kanaverse/kana#237 @llewelld
Thanks for including this (and for your gracious credit); these changes address the issue for me.
