### 2023-07-31 kanaverse/kana#238 @llewelld
The wiki is a nice start. This kind of documentation would be a great place to start fostering a community as it's a low-bar entry point for new contributors. Currently there doesn't seem to be an easy way for non-project-members to contribute to the wiki.

The reason I noticed is because I was hoping to create a PR to update a link in Home.md, which seems to be misdirected (or maybe missing?) right now.

*How to use the explore mode*
From: https://github.com/kanaverse/kana/wiki/How-to-Explore-mode
To: https://github.com/kanaverse/kana/wiki/How-to-explore-pre%E2%80%90analyzed-single%E2%80%90cell-datasets
