### 2023-07-31 kanaverse/kana#238 @yongrenjie
Looks good! The only thing I'd suggest is to add a link to this wiki to the kana web app itself, so that it's more discoverable. (Unless there's already one and I missed it?)
