### 2023-08-01 kanaverse/kana#234 @LTLA
That's correct. For the public test datasets, we're talking about 10 - 500 MB. (If interested, you can inspect the sizes at https://github.com/kanaverse/random-test-files/releases.) Internally, we have see kana used with files up to 1 GB in size. The results themselves are usually several MB, probably not more than 10 MB in total but split over multiple requests while the user interacts with the application. This is a less obvious source of sluggishness, e.g., when the app has to constantly query the backend for expression values when the user picks a different gene for coloring the t-SNE/UMAP.

I added some ballpark figures to the manuscript, fleshed out in kanaverse/kana-paper#10.

> Biology is not our specific area of expertise, but we wonder whether, for a biologist working in bioinformatics, an R-based solution wouldn't nevertheless work well for them (we expect R to be a tool many biologists would be familiar with). Given this, would it be possible to more fully justify the use of the browser and the impacts of doing so in the text?

FWIW I think that expectation is... optimistic, at least in my personal experience with collaborating with bench scientists.
