### 2023-08-01 kanaverse/kana#232 @LTLA
> Is there a reason you don't cite your bioRxiv paper directly? This would seem like a natural thing to do and provide an easy way for the reader to drill down further into the detail.

Oh. Are we allowed to do that? Seems kinda recursive to cite a different version of the same paper.

I'm happy to do it. Just didn't know whether it was "proper".
