### 2023-08-02 77af7a4
Fleshed out user advantages of client-side compute.
