### 2023-08-02 kanaverse/kana#234 @llewelld
Thanks for making these changes; these look good to me, and I think make a much clearer and stronger case for the use of client-side compute. @yongrenjie, what do you think?
