### 2023-08-02 kanaverse/kana#232 @AoifeHughes
I **think** this all seems fine. It would be great if you add the reference to the bioRxiv paper.

From a quick skim of the current draft of the paper I would suggest that the `Further comments` section be reduced and the above mentioned table be added. Having the concrete numbers for performance values is very useful for a technical audience.
