### 2023-08-02 24aa9e7
Minor wording tweaks for the Contributing file (#240)
