+++ README.md
@@ -52,6 +52,8 @@ The interface provides a depiction of the dimensionality reduction of choice,
 a ranking of marker genes for the cluster of interest,
 and diagnostic plots from the individual analysis steps.
 
+***Checkout the [wiki](https://github.com/kanaverse/kana/wiki) for tutorials on the functionality Kana provides.***
+
 ![Features](assets/v3_release.png)
 
 **Tips and tricks:**
