### 2023-08-02 kanaverse/kana#238 @jkanche
I've fixed the link to explore datasets and also added the link to the wiki in the README.
