### 2023-08-03 ce1d735
Added the table and citation back to self.
