+++ ref.bib
@@ -678,3 +678,15 @@ url = {https://CRAN.R-project.org/package=shiny},
    Month="Apr",
    doi="10.1186/s13059-016-0947-7"
 }
+
+@article{otherkana,
+    author = {Aaron Lun and Jayaram Kancherla},
+    title = {Powering single-cell analyses in the browser with WebAssembly},
+    elocation-id = {2022.03.02.482701},
+    year = {2023},
+    doi = {10.1101/2022.03.02.482701},
+    publisher = {Cold Spring Harbor Laboratory},
+    URL = {https://www.biorxiv.org/content/early/2023/05/15/2022.03.02.482701},
+    eprint = {https://www.biorxiv.org/content/early/2023/05/15/2022.03.02.482701.full.pdf},
+    journal = {bioRxiv}
+}
