### 2023-08-03 kanaverse/kana#232 @LTLA
Alright, as suggested, I added an abbreviated version of the table, added the citation to the bioRxiv paper,  and trimmed out some stuff from the Further comments (though not too much, in order to still address #234 properly).

Changes are in kanaverse/kana-paper#11; I'll trigger a new build on the main issue.
