### 2023-08-04 kanaverse/kana#232 @llewelld
Personally, I really like this. I think the data is a worthwhile addition and I agree with @AoifeHughes that this is likely to be of interest to readers. I defer to @AoifeHughes and @yongrenjie, but as far as I'm concerned this addresses the original issue (it actually goes beyond what we asked for, but in a very positive way).
