### 2023-08-10 #5603 @llewelld
A quick recap on where things stand as I see it. All of the issues now look good to me. I think we are still waiting on a final decision in relation to #232. @yongrenjie and I will have to go through the [checklist above](https://github.com/openjournals/joss-reviews/issues/5603#issuecomment-1612688573) again; it looks like everything should be resolved there already, but we should still complete the process as a formality. There may be a slight delay before I can do this (my apologies) since we hit August with vacations.

@yongrenjie: would you be available to take a look at this together next week?
