### 2023-08-16 #5603 @yongrenjie
It seems like it should be all good — the only outstanding issue is https://github.com/kanaverse/kana/issues/232, where we're just waiting for confirmation (though my opinion is that it shouldn't be a problem and we can probably close it soon).

I've ticked off the items in the checklist for now.

@AoifeHughes Is there anything else we need to do as reviewers, apart from recommending that the paper be accepted in its current form 😉?
