### 2023-08-22 #5603 @jkanche
@AoifeHughes We already have a tagged v3.0.0 release of Kana and made the submission on zenodo: https://zenodo.org/record/8273506

Let me know if this works for you.
