### 2023-08-22 #5603 @jkanche
Tagged a new release (v3.0.1) and the [zenodo link](https://zenodo.org/record/8273645) should reflect these changes.
