### 2023-09-04 #5603 @Kevin-Mattheus-Moerman
@LTLA I am the AEiC for this track and here to help with final steps. I have check this review, your repository, the archive, and the paper. Most seems in order. I only have the below points that require your attention:

- [ ] Please edit your archive listed software license to match your actual software license. It currently says ` Creative Commons Attribution 4.0 International` but it should refer to your `MIT license`.
- [ ] In your affiliations please spell out `USA` as `United States of America`.
