+++ ref.bib
@@ -359,7 +359,9 @@ url = {https://igraph.org},
   number={6226},
   pages={1138--1142},
   year={2015},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.aaa1934}
+
 }
 
 @article{paul2015transcriptional,
@@ -370,7 +372,8 @@ url = {https://igraph.org},
   number={7},
   pages={1663--1677},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cell.2015.11.013}
 }
 
 @article{bach2017differentiation,
@@ -425,7 +428,8 @@ url = {https://igraph.org},
   author={Zakai, Alon},
   booktitle={Proceedings of the ACM international conference companion on Object oriented programming systems languages and applications companion},
   pages={301--312},
-  year={2011}
+  year={2011},
+  doi={10.1145/2048147.2048224}
 }
 
 @inproceedings {jangda2019not,
